+++ packages/ui/docs/icloud/stories/table/index.stories.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import Table from '@osui/table';
+
+export default {
+    title: 'Table',
+};
+
+export const Demo = () => {
+
+    const columns = [
+        {
+            title: 'Name',
+            dataIndex: 'name',
+            key: 'name',
+        },
+        {
+            title: 'Age',
+            dataIndex: 'age',
+            key: 'age',
+        },
+        {
+            title: 'Address',
+            dataIndex: 'address',
+            key: 'address',
+        },
+    ];
+
+    const data = [
+        {
+            key: '1',
+            name: 'John Brown',
+            age: 32,
+            address: 'New York No. 1 Lake Park',
+        },
+        {
+            key: '2',
+            name: 'Jim Green',
+            age: 42,
+            address: 'London No. 1 Lake Park',
+        },
+        {
+            key: '3',
+            name: 'Joe Black',
+            age: 32,
+            address: 'Sidney No. 1 Lake Park',
+        },
+    ];
+
+
+    return (
+        <div style={{padding: 30}}>
+            <Table columns={columns} dataSource={data} />
+        </div>);
+};