+++ packages/ui/docs/osui/stories/markdown/index.stories.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import Markdown from '@osui/markdown';
+
+export default {
+    title: '业务/Markdown',
+};
+
+const content = `
+# Writing template strings in Markdown
+
+## Writing template strings in Markdown
+
+adsfasdfasdf
+
+asdfasdf
+
+\`\`\`jsx
+<Markdown content={content} />
+\`\`\`
+`;
+
+export const Demo = () => (
+    <Markdown content={content} />
+);