+++ packages/ui-icons/osui-icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://gitee.com/gitee-fe/osui/compare/@osui/icons@0.4.3...@osui/icons@0.5.0) (2020-09-14)
+
+**Note:** Version bump only for package @osui/icons
+
+
+
+
+
 ## [0.4.3](https://gitee.com/gitee-fe/osui/compare/@osui/icons@0.4.2...@osui/icons@0.4.3) (2020-09-08)
 
 **Note:** Version bump only for package @osui/icons